🎨: 장바구니 총 금액 계산 로직을 renderCalculateCart 함수로 추출
